+++ lib/members_page.rb
@@ -1,0 +1,12 @@
+# frozen_string_literal: true
+require 'scraped'
+
+class MembersPage < Scraped::HTML
+  decorator Scraped::Response::Decorator::AbsoluteUrls
+
+  field :members do
+    noko.css('section.article-content div#o1').map do |mem|
+      fragment mem => MemberDiv
+    end
+  end
+end
